Changes to get examples going.

+++ src/python/monodomain_2D_vary_parameters_spatially.py
@@ -274,7 +274,7 @@ lastNodeDomain = decomposition.NodeDomainGet(lastNodeNumber, 1)
 
 # Set the stimulus on half the bottom nodes
 stimComponent = cellML.FieldComponentGet(cellModel, iron.CellMLFieldTypes.PARAMETERS, "membrane/IStim")
-for node in range(1,numberOfXElements/2):
+for node in range(1,int(numberOfXElements/2)):
     nodeDomain = decomposition.NodeDomainGet(node,1)
     if nodeDomain == computationalNodeNumber:
         cellMLParametersField.ParameterSetUpdateNode(iron.FieldVariableTypes.U, iron.FieldParameterSetTypes.VALUES, 1, 1, node, stimComponent, stimValue)
@@ -348,7 +348,7 @@ solverEquations.BoundaryConditionsCreateFinish()
 problem.Solve()
 
 # Now turn the stimulus off
-for node in range(1,numberOfXElements/2):
+for node in range(1,int(numberOfXElements/2)):
     nodeDomain = decomposition.NodeDomainGet(node,1)
     if nodeDomain == computationalNodeNumber:
         cellMLParametersField.ParameterSetUpdateNode(iron.FieldVariableTypes.U, iron.FieldParameterSetTypes.VALUES, 1, 1, node, stimComponent, 0.0)
